Effect of smartphone applications on disease control report
Simulating the Effects of Social-Distancing Technology on the Spread of a Pandemic

Our research currently match the empirical studies done on testing the effectiveness of a contact tracing application in a population by Oxford University. As
the results suggested that if more than half the population use this application with the required social distancing interventions, it has the potential to stop the spread of the Coronavirus.